Move decimal constants from `arrow-data` to `arrow-schema` crate
